requirements.txt
kaleido==0.2.1
loguru==0.7.0
numpy==1.25.0
packaging==23.1
pandas==2.0.3
patsy==0.5.3
plotly==5.15.0
plotly-express==0.4.1
plotly-utils==0.0.3
python-dateutil==2.8.2
pytz==2023.3
request==2.28.1
scipy==1.11.1
six==1.16.0
statsmodels==0.14.0
tenacity==8.2.2
tzdata==2023.3

